backend/pyproject.toml
[project]
name = "my_ledge-backend"
version = "0.1.0"
description = "Backend service for my_ledge"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13,<2.0",
  "msoffcrypto-tool>=5.4,<6.0",
  "fastapi>=0.115,<1.0",
  "openpyxl>=3.1,<4.0",
  "pydantic-settings>=2.7,<3.0",
  "sqlalchemy>=2.0,<3.0",
  "asyncpg>=0.30,<1.0",
  "python-multipart>=0.0.9,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[dependency-groups]
dev = [
  "aiosqlite>=0.20,<1.0",
  "httpx>=0.27,<1.0",
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.24,<1.0",
  "ruff>=0.6,<1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88
